# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TheraDOS \u2014 Therapeutic Domain Operating System",
    "description": "Provenance-aware therapeutic intelligence operating system converting heterogeneous evidence into falsifiable, safety-constrained hypotheses.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/evidence": {
      "get": {
        "tags": [
          "Evidence & Provenance"
        ],
        "summary": "List Evidence",
        "operationId": "list_evidence_api_v1_evidence_get",
        "parameters": [
          {
            "name": "include_synthetic",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Set to True to include tutorial synthetic evidence",
              "default": false,
              "title": "Include Synthetic"
            },
            "description": "Set to True to include tutorial synthetic evidence"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceRecordRead"
                  },
                  "title": "Response List Evidence Api V1 Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/ingest": {
      "post": {
        "tags": [
          "Evidence & Provenance"
        ],
        "summary": "Ingest Evidence",
        "operationId": "ingest_evidence_api_v1_evidence_ingest_post",
        "parameters": [
          {
            "name": "is_tutorial_project",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Set to True only if ingesting into a synthetic tutorial project",
              "default": false,
              "title": "Is Tutorial Project"
            },
            "description": "Set to True only if ingesting into a synthetic tutorial project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceRecordCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hypotheses": {
      "get": {
        "tags": [
          "Therapeutic Hypotheses"
        ],
        "summary": "List Hypotheses",
        "operationId": "list_hypotheses_api_v1_hypotheses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HypothesisRead"
                  },
                  "type": "array",
                  "title": "Response List Hypotheses Api V1 Hypotheses Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Therapeutic Hypotheses"
        ],
        "summary": "Create Hypothesis",
        "operationId": "create_hypothesis_api_v1_hypotheses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HypothesisCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HypothesisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hypotheses/{hypothesis_id}/compile": {
      "post": {
        "tags": [
          "Therapeutic Hypotheses"
        ],
        "summary": "Compile Hypothesis",
        "operationId": "compile_hypothesis_api_v1_hypotheses__hypothesis_id__compile_post",
        "parameters": [
          {
            "name": "hypothesis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hypothesis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Compile Hypothesis Api V1 Hypotheses  Hypothesis Id  Compile Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hypotheses/{hypothesis_id}/falsify": {
      "post": {
        "tags": [
          "Therapeutic Hypotheses"
        ],
        "summary": "Falsify Hypothesis",
        "operationId": "falsify_hypothesis_api_v1_hypotheses__hypothesis_id__falsify_post",
        "parameters": [
          {
            "name": "hypothesis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hypothesis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Falsify Hypothesis Api V1 Hypotheses  Hypothesis Id  Falsify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlternativeMechanismRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "mechanism_name": {
            "type": "string",
            "title": "Mechanism Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "evidence_support": {
            "type": "number",
            "title": "Evidence Support"
          },
          "discriminating_assay": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discriminating Assay"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mechanism_name",
          "description",
          "evidence_support"
        ],
        "title": "AlternativeMechanismRead"
      },
      "Body_login_api_v1_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_v1_auth_login_post"
      },
...
      "EvidenceMaturity": {
        "type": "string",
        "enum": [
          "associative",
          "structurally_inferred",
          "mechanistically_supported",
          "causally_corroborated",
          "experimentally_validated",
          "clinically_established"
        ],
        "title": "EvidenceMaturity"
      },
      "EvidenceRecordCreate": {
        "properties": {
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "evidence_type": {
            "type": "string",
            "title": "Evidence Type"
          },
          "evidence_maturity": {
            "$ref": "#/components/schemas/EvidenceMaturity",
            "default": "associative"
          },
          "normalized_payload": {
            "type": "object",
            "title": "Normalized Payload"
          },
          "quality_score": {
            "type": "number",
            "title": "Quality Score",
            "default": 1.0
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "source_name",
          "evidence_type",
          "normalized_payload"
        ],
        "title": "EvidenceRecordCreate"
      },
      "EvidenceRecordRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "evidence_type": {
            "type": "string",
            "title": "Evidence Type"
          },
          "evidence_maturity": {
            "$ref": "#/components/schemas/EvidenceMaturity"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum"
          },
          "quality_score": {
            "type": "number",
            "title": "Quality Score"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic"
          },
          "retrieval_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieval Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "evidence_type",
          "evidence_maturity",
          "checksum",
          "quality_score",
          "is_synthetic",
          "retrieval_timestamp"
        ],
        "title": "EvidenceRecordRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HypothesisCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "intervention_name": {
            "type": "string",
            "title": "Intervention Name"
          },
          "intended_target": {
            "type": "string",
            "title": "Intended Target"
          },
          "intended_action": {
            "type": "string",
            "title": "Intended Action"
          },
          "cellular_context": {
            "type": "string",
            "title": "Cellular Context"
          },
          "disease_endotype": {
            "type": "string",
            "title": "Disease Endotype"
          },
          "genomic_background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genomic Background"
          },
          "predictive_biomarkers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Predictive Biomarkers"
          },
          "dose_exposure_regime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dose Exposure Regime"
          },
          "schedule_duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Duration"
          },
          "safety_constraints": {
            "type": "object",
            "title": "Safety Constraints"
          },
          "program_id": {
            "type": "string",
            "title": "Program Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "intervention_name",
          "intended_target",
          "intended_action",
          "cellular_context",
          "disease_endotype",
          "program_id"
        ],
        "title": "HypothesisCreate"
      },
      "HypothesisRead": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "intervention_name": {
            "type": "string",
            "title": "Intervention Name"
          },
          "intended_target": {
            "type": "string",
            "title": "Intended Target"
          },
          "intended_action": {
            "type": "string",
            "title": "Intended Action"
          },
          "cellular_context": {
            "type": "string",
            "title": "Cellular Context"
          },
          "disease_endotype": {
            "type": "string",
            "title": "Disease Endotype"
          },
          "genomic_background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genomic Background"
          },
          "predictive_biomarkers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Predictive Biomarkers"
          },
          "dose_exposure_regime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dose Exposure Regime"
          },
          "schedule_duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Duration"
          },
          "safety_constraints": {
            "type": "object",
            "title": "Safety Constraints"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "program_id": {
            "type": "string",
            "title": "Program Id"
          },
          "maturity": {
            "$ref": "#/components/schemas/EvidenceMaturity"
          },
          "support_score": {
            "type": "number",
            "title": "Support Score"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "proof_obligations": {
            "items": {
              "$ref": "#/components/schemas/ProofObligationRead"
            },
            "type": "array",
            "title": "Proof Obligations"
          },
          "alternative_mechanisms": {
            "items": {
              "$ref": "#/components/schemas/AlternativeMechanismRead"
            },
            "type": "array",
            "title": "Alternative Mechanisms"
          }
        },
        "type": "object",
        "required": [
          "title",
          "intervention_name",
          "intended_target",
          "intended_action",
          "cellular_context",
          "disease_endotype",
          "id",
          "program_id",
          "maturity",
          "support_score",
          "status",
          "version",
          "created_at"
        ],
        "title": "HypothesisRead"
      },
...
      "ProofObligationRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "proposition": {
            "type": "string",
            "title": "Proposition"
          },
          "required_evidence_type": {
            "type": "string",
            "title": "Required Evidence Type"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "threshold_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold Value"
          },
          "evidence_references": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence References"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "proposition",
          "required_evidence_type",
          "state",
          "evidence_references"
        ],
        "title": "ProofObligationRead"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "Token"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "display_name",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserRead": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "display_name",
          "id",
          "status",
          "created_at"
        ],
        "title": "UserRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}